@@ -5,22 +5,22 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1.0",
         "alerj/alerj-auth": "^1.0",
         "asantibanez/livewire-calendar": "^2.2",
         "asantibanez/livewire-charts": "^2.5",
         "barryvdh/laravel-dompdf": "^2.0",
         "bugsnag/bugsnag-laravel": "^2.23",
-        "doctrine/dbal": "^3.3",
+        "doctrine/dbal": "^3.0",
         "endroid/qr-code": "^4.8",
         "geekcom/validator-docs": "^3.7",
         "google/cloud-recaptcha-enterprise": "^1.7",
         "guzzlehttp/guzzle": "^7.2",
         "http-interop/http-factory-guzzle": "^1.2",
-        "laravel/framework": "^9.2",
+        "laravel/framework": "^10.0",
         "laravel/horizon": "^5.21",
         "laravel/jetstream": "^2.7",
-        "laravel/sanctum": "^2.14.1",
+        "laravel/sanctum": "^3.2",
         "laravel/scout": "^10.8",
         "laravel/tinker": "^2.7",
         "livewire/livewire": "^2.10",
@@ -38,10 +38,9 @@
         "laravel/dusk": "^7.11",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
-        "nascent-africa/jetstrap": "^2.5",
-        "nunomaduro/collision": "^6.1",
-        "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.0"
+        "nunomaduro/collision": "^7.0",
+        "phpunit/phpunit": "^10.0",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -94,6 +93,6 @@
             "url": "git@github.com:alerj/alerj-auth.git"
         }
     ],
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
